Fix FindSCC macro
SCC_FOUND is not generated by find_program according to the cmake
documentation. No idea why this ever worked. Maybe in older versions of
cmake this was the case. In any case, just checking SCC_EXECUTABLE is
sufficient since this variable will be set to a value that evaluates to
false in case SCC cannot be found.