hasSignalArrived -> lastArrivedSignal
Make it more easy for users to understand how to retrieve the last
arrived signal in working loops. Instead of requiring them to use
waitForSignal in the assumption that it won't block anymore, let the
non-blocking method immediately return the signal that arrived or a
special enum entry in case no signal arrived.

* examples/SignalWaiterLoop.cpp: new example for loops
* src/rsc/misc/MacSignalWaiter.cpp: adapted to new API
* src/rsc/misc/PosixSignalWaiter.cpp: likewise
* src/rsc/misc/SignalWaiter.h: changed API and improved comments
* src/rsc/misc/Win32SignalWaiter.cpp: adapted to new API