plugin configurator: NoSuchObject if plugin not found
Throw an exception in the plugin configurator in case a requested plugin
was not found. This behavior can be switched with a boolean in a new
method that has been created to actually execute the configuration of
the plugin system.

This commit breaks the existing API and requires clients to call
execute() on a used Configurator object. This change has been made to
prevent throwing exceptions in destructors.

* src/rsc/plugins/Configurator.cpp: moved destructor code to execute
  method + new exception in case a plugin is missing if requested.
* src/rsc/plugins/Configurator.h: likewise
* test/rsc/plugins/ConfiguratorTest.cpp: test for the new behavior

fixes #2488